+++ stream-chat-android-client/src/debug/java/io/getstream/chat/android/client/di/ChatModule.kt
@@ -19,6 +19,7 @@ package io.getstream.chat.android.client.di
 import android.content.Context
 import androidx.lifecycle.Lifecycle
 import io.getstream.chat.android.client.api.ChatClientConfig
+import io.getstream.chat.android.client.debugger.ChatClientDebugger
 import io.getstream.chat.android.client.notifications.handler.NotificationConfig
 import io.getstream.chat.android.client.notifications.handler.NotificationHandler
 import io.getstream.chat.android.client.parser.ChatParser
@@ -44,6 +45,7 @@ internal class ChatModule(
     uploader: FileUploader?,
     tokenManager: TokenManager,
     customOkHttpClient: OkHttpClient?,
+    clientDebugger: ChatClientDebugger?,
     lifecycle: Lifecycle,
 ) : BaseChatModule(
     appContext,
@@ -55,6 +57,7 @@ internal class ChatModule(
     uploader,
     tokenManager,
     customOkHttpClient,
+    clientDebugger,
     lifecycle,
 ) {