+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -3129,6 +3129,7 @@ internal constructor(
                     fileUploader,
                     tokenManager,
                     customOkHttpClient,
+                    clientDebugger,
                     lifecycle,
                 )